# backend/models/user.go
package models

import (
	"time"
)

type User struct {
	ID        uint       `json:"id" gorm:"primaryKey"`
	CreatedAt time.Time  `json:"created_at"`
	UpdatedAt time.Time  `json:"updated_at"`
	DeletedAt *time.Time `json:"deleted_at,omitempty"`
	Username  string     `json:"username" gorm:"unique;not null"`
	Password  string     `json:"password"`
}